+++ src/features/breather/index.js
@@ -10,10 +10,23 @@ import {registerBreatherSettings} from './settings';
 import {Breather} from './breather';
 import {registerBreatherTests} from './quench';
 
+function registerBreatherRestType() {
+    CONFIG.DND5E.restTypes.breather = {
+        duration: {
+            normal: 5,
+            gritty: 5,
+            epic: 5
+        },
+        label: 'sosly.breather.title',
+        icon: 'fa-solid fa-face-exhaling'
+    };
+}
+
 export function registerBreatherFeature() {
     logger.info('Registering Breather Rest System');
 
     registerBreatherSettings();
+    registerBreatherRestType();
 
     if (game.settings.get(module_id, 'breather')) {
         Breather.register();